+++ NEWS.md
@@ -1,3 +1,8 @@
+DALEX 2.0.1
+---------------------------------------------------------------
+* Fixed R package not working with Python Explainer ([#318](https://github.com/ModelOriented/DALEX/issues/318))
+* Fixed `model_diagnostics()` returning an error when `y_hat` or `residuals` is of `array` class ([#319](https://github.com/ModelOriented/DALEX/issues/319))
+
 DALEX 2.0
 ---------------------------------------------------------------
 * `plot.predict_diagnostics` now passess ellipsis to `plot.ceteris_paribus_explainer`